\name{PredictL1Package-package}
\alias{PredictL1Package-package}
\alias{PredictL1Package}
\docType{package}
\title{
\packageTitle{PredictL1Package}
}
\description{
\packageDescription{PredictL1Package}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PredictL1Package}
\packageIndices{PredictL1Package}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PredictL1Package}

Maintainer: \packageMaintainer{PredictL1Package}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
